requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
langgraph
chromadb
sentence-transformers
python-dotenv
Pillow
easyocr
numpy
soundfile
openai-whisper
pydub

langchain-groq
streamlit-audiorecorder
